renamed files to standard names
